Move env controllers into own file
